[INFRA] Move core/algorithm into core/configuration
algorithm/concept.hpp -> core/configuration/detail/.
algorithm/configuration.hpp -> core/configuration/.
algorithm/configuration_element_debug_mode.hpp -> core/configuration/detail/.
algorithm/configuration_element_parallel_mode.hpp -> core/configuration/detail/.
algorithm/configuration_utility.hpp -> core/configuration/detail/.

Skipped: algorithm/pipeable_config_element.hpp (needs a decision)

Keep untouched:
algorithm/detail/... -> core/algorithm/detail/...
algorithm/algorithm_result_generator_range -> core/algorithm/algorithm_result_generator_range

Added new all.hpp's and adjusted the other ones.
