+++ include/seqan3/alignment/configuration/detail.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/core/algorithm/configuration_utility.hpp>
+#include <seqan3/core/configuration/detail/configuration_utility.hpp>
 
 namespace seqan3::detail
 {
