+++ include/seqan3/alignment/pairwise/alignment_result.hpp
@@ -15,7 +15,7 @@
 
 #include <optional>
 
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
