+++ include/seqan3/alignment/pairwise/detail/policy_optimum_tracker.hpp
@@ -18,8 +18,8 @@
 #include <seqan3/alignment/matrix/detail/coordinate_matrix.hpp>
 #include <seqan3/alignment/matrix/detail/matrix_coordinate.hpp>
 #include <seqan3/alignment/pairwise/detail/type_traits.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/concept/tuple.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3::detail
