+++ include/seqan3/alignment/pairwise/policy/simd_affine_gap_policy.hpp
@@ -19,8 +19,8 @@
 #include <seqan3/alignment/matrix/alignment_optimum.hpp>
 #include <seqan3/alignment/matrix/trace_directions.hpp>
 #include <seqan3/alignment/pairwise/detail/alignment_algorithm_state.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/simd/concept.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
