+++ include/seqan3/core/algorithm/pipeable_config_element.hpp
@@ -16,7 +16,7 @@
 #include <tuple>
 #include <type_traits>
 
-#include <seqan3/core/algorithm/concept.hpp>
+#include <seqan3/core/configuration/detail/concept.hpp>
 #include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3
