+++ include/seqan3/core/all.hpp
@@ -17,9 +17,10 @@
 // External concept implementations
 // ============================================================================
 
-#include <seqan3/core/algorithm/all.hpp>
 #include <seqan3/core/add_enum_bitwise_operators.hpp>
+#include <seqan3/core/algorithm/all.hpp>
 #include <seqan3/core/concept/all.hpp>
+#include <seqan3/core/configuration/all.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/core/detail/all.hpp>
 #include <seqan3/core/platform.hpp>
