+++ include/seqan3/search/configuration/detail.hpp
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include <seqan3/core/algorithm/configuration_utility.hpp>
+#include <seqan3/core/configuration/detail/configuration_utility.hpp>
 
 namespace seqan3::detail
 {
