+++ include/seqan3/search/search.hpp
@@ -16,8 +16,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/algorithm/algorithm_result_generator_range.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/algorithm/detail/algorithm_executor_blocking.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/range/views/convert.hpp>
 #include <seqan3/range/views/deep.hpp>
 #include <seqan3/range/views/persist.hpp>
