+++ include/seqan3/search/search_result.hpp
@@ -16,7 +16,7 @@
 #include <seqan3/std/concepts>
 #include <exception>
 
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
