+++ test/snippet/alignment/configuration/align_cfg_score_type.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/alignment/configuration/align_config_score_type.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 
 int main()
 {
