+++ test/snippet/core/configuration/configuration_combine.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alignment/configuration/align_config_band.hpp>
 #include <seqan3/alignment/configuration/align_config_method.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 
 int main()
 {
