+++ test/snippet/core/configuration/configuration_compatibility.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 
 enum struct my_id : int
 {
