+++ test/snippet/core/configuration/configuration_get_or.cpp
@@ -1,5 +1,5 @@
-#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/core/debug_stream.hpp>
 
 // Initial setup used in the actual example:
