+++ test/snippet/search/configuration_error.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/search/configuration/max_error.hpp>
 
 int main()
