+++ test/snippet/search/configuration_parallel.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/search/configuration/max_error.hpp>
 #include <seqan3/search/configuration/parallel.hpp>
