+++ test/unit/core/algorithm/CMakeLists.txt
@@ -1,6 +1,5 @@
 seqan3_test(algorithm_result_generator_range_test.cpp)
 seqan3_test (bound_test.cpp)
-seqan3_test (configuration_test.cpp)
 seqan3_test (pipeable_config_element_test.cpp)
 
 add_subdirectories()
