+++ test/unit/core/configuration/CMakeLists.txt
@@ -1,0 +1,3 @@
+seqan3_test (configuration_test.cpp)
+
+add_subdirectories()
